requirements.txt
openai==2.5.0
python-dotenv==1.1.1
mcp==1.18.0
neo4j==6.0.2
neo4j-graphrag==1.10.0